[RA] Fixes crash on 64bit linux builds.
Fixes crash when a bib is removed.
Alters TD code to be the same for consistency.

+++ redalert/defines.h
@@ -492,7 +492,7 @@ typedef union
     } Sub;
 } LEPTON_COMPOSITE;
 
-typedef unsigned long COORDINATE;
+typedef unsigned COORDINATE;
 typedef union
 {
     COORDINATE Coord;
